Fix nav_graph long argType crash: change mealLogId to integer
Navigation Component XML inflater cannot represent long literals —
it parses any numeric defaultValue as integer, causing
XmlPullParserException. Changed mealLogId argType from long to integer
and updated Bundle put/get calls to match (putInt/getInt).

+++ app/src/main/java/com/pablosanz/gymapp/ui/nutrition/MealDetailFragment.java
@@ -48,7 +48,7 @@ public class MealDetailFragment extends Fragment {
         if (getArguments() != null) {
             mealSlot = getArguments().getString("mealSlot", "desayuno");
             date = getArguments().getString("date", DateUtils.today());
-            mealLogId = getArguments().getLong("mealLogId", -1L);
+            mealLogId = getArguments().getInt("mealLogId", -1);
         }
 
         repo = new NutritionRepository(requireActivity().getApplication());
